move test_filter_on_death_date to pre-existing test file
